Update token for tests
