+++ test/ReportPortal.Client.IntegrationTests/LaunchItem/LaunchItemFixture.cs
@@ -13,80 +13,6 @@ namespace ReportPortal.Client.IntegrationTests.LaunchItem
     public class LaunchItemFixture : BaseFixture
     {
         [Fact]
-        public async Task GetInvalidLaunch()
-        {
-            await Assert.ThrowsAsync<ReportPortalException>(async () => await Service.Launch.GetAsync("invalid_id"));
-        }
-
-        [Fact]
-        public async Task GetLaunches()
-        {
-            var container = await Service.Launch.GetAsync();
-            var launches = container.Items.ToList();
-            Assert.True(launches.Count() > 0);
-        }
-
-        [Fact]
-        public async Task GetDebugLaunches()
-        {
-            var launches = await Service.Launch.GetDebugAsync();
-            launches.Items.ToList().ForEach((l) => Assert.Equal(LaunchMode.Debug, l.Mode));
-        }
-
-        [Fact]
-        public async Task GetTheFirst10Launches()
-        {
-            var launches = await Service.Launch.GetAsync(new FilterOption
-            {
-                Paging = new Paging(1, 10)
-            });
-            Assert.Equal(10, launches.Items.Count());
-        }
-
-        [Fact]
-        public async Task GetLaunchesFilteredByName()
-        {
-            var launches = await Service.Launch.GetAsync(new FilterOption
-            {
-                Paging = new Paging(1, 10),
-                Filters = new List<Filter> { new Filter(FilterOperation.Contains, "name", "test") }
-            });
-            Assert.True(launches.Items.Count() > 0);
-            foreach (var launch in launches.Items)
-            {
-                Assert.Contains("test", launch.Name.ToLower());
-            }
-        }
-
-        [Fact]
-        public async Task GetLaunchesSortedByAscendingDate()
-        {
-            var launches = await Service.Launch.GetAsync(new FilterOption
-            {
-                Paging = new Paging(1, 10),
-                Sorting = new Sorting(new List<string> { "startTime" }, SortDirection.Ascending)
-            });
-
-            Assert.True(launches.Items.Count() > 0);
-
-            Assert.Equal(launches.Items.Select(l => l.StartTime).OrderBy(st => st), launches.Items.Select(l => l.StartTime));
-        }
-
-        [Fact]
-        public async Task GetLaunchesSortedByDescendingDate()
-        {
-            var launches = await Service.Launch.GetAsync(new FilterOption
-            {
-                Paging = new Paging(1, 10),
-                Sorting = new Sorting(new List<string> { "startTime" }, SortDirection.Descending)
-            });
-
-            Assert.True(launches.Items.Count() > 0);
-
-            Assert.Equal(launches.Items.Select(l => l.StartTime).OrderByDescending(st => st), launches.Items.Select(l => l.StartTime));
-        }
-
-        [Fact]
         public async Task StartFinishDeleteLaunch()
         {
             var startLaunchRequest = new StartLaunchRequest
@@ -378,30 +304,6 @@ namespace ReportPortal.Client.IntegrationTests.LaunchItem
             });
 
             var getLaunch = await Service.Launch.GetAsync(launch.Uuid);
-
-            var delMessage = await Service.Launch.DeleteAsync(getLaunch.Id);
-            Assert.Contains("successfully", delMessage.Info);
-        }
-
-        [Fact]
-        public async Task GetInProgressLaunch()
-        {
-            var launch = await Service.Launch.StartAsync(new StartLaunchRequest
-            {
-                Name = "StartForceFinishIncompleteLaunch",
-                StartTime = DateTime.UtcNow,
-                Mode = LaunchMode.Default
-            });
-
-            var getLaunch = await Service.Launch.GetAsync(launch.Uuid);
-
-            Assert.NotEqual(default(DateTime), getLaunch.StartTime);
-            Assert.Null(getLaunch.EndTime);
-
-            await Service.Launch.FinishAsync(launch.Uuid, new FinishLaunchRequest
-            {
-                EndTime = DateTime.UtcNow
-            });
 
             var delMessage = await Service.Launch.DeleteAsync(getLaunch.Id);
             Assert.Contains("successfully", delMessage.Info);
