+++ test/ReportPortal.Client.IntegrationTests/User/UserFixture.cs
@@ -11,8 +11,8 @@ namespace ReportPortal.Client.IntegrationTests.User
         {
             var user = await Service.User.GetAsync();
             // Assert.Equal("Used for Net integration check via CI", user.Fullname);
-            //Assert.Equal("tester", user.Fullname);
-            Assert.Equal("default", user.Fullname);
+            Assert.Equal("tester", user.Fullname);
+            //Assert.Equal("default", user.Fullname);
             Assert.NotEmpty(user.Email);
 
             Assert.NotNull(user.AssignedProjects);
